Modify CSS build dependency & update CSS builds
In superproject's CSS build depdendency ucore-custom.less, add a default fade-in animation to sectional article headers. Next, update the submodule for the DAESA-wide CSS development project to incorporate changes in the dependencies it contains into the superproject. Finally, rebuild the website's custom CSS production files to integrate all of these dependency changes into the final stylesheet that is applied to the site via WSUWP.